+++ Changelog.md
@@ -1,6 +1,13 @@
 # Project IceForest Calculator C
 ### Changelog
 
+## Version 1.2.x
+### Version 1.2 (March 7, 2019)
+* Added a Pythagorean Triple Finder.
+* Minor change regarding the titles for some of the calculation options.
+* Updated the Quadratic Equation Solver.
+* Other minor change(s).
+
 ## Version 1.1.x
 ### Version 1.1.1 (February 12, 2019)
 * Minor change in Exponents.